Finished refactoring Papyrus Script classes.
Moved MakeSPSnippetPromise to a proper place.
