+++ skymp5-server/cpp/server_guest_lib/script_classes/IPapyrusClass.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "SpSnippetFunctionGen.h"
 #include "papyrus-vm/VirtualMachine.h"
 #include "script_compatibility_policies/IPapyrusCompatibilityPolicy.h"
