+++ skymp5-server/cpp/server_guest_lib/script_classes/PapyrusCell.cpp
@@ -1,5 +1,6 @@
 #include "PapyrusCell.h"
 #include "script_objects/EspmGameObject.h"
+#include "spdlog/spdlog.h"
 
 VarValue PapyrusCell::IsAttached(VarValue self,
                                  const std::vector<VarValue>& arguments)
