+++ skymp5-server/cpp/server_guest_lib/script_classes/PapyrusLeveledBase.cpp
@@ -39,3 +39,8 @@ void PapyrusLeveledBase::Register(
 
   AddMethod(vm, "GetNthForm", &PapyrusLeveledBase::GetNthForm);
 }
+
+PapyrusLeveledBase::PapyrusLeveledBase(const std::string& name)
+  : strName(name)
+{
+}
