+++ skymp5-server/cpp/server_guest_lib/script_classes/PapyrusLeveledBase.h
@@ -4,10 +4,7 @@
 class PapyrusLeveledBase : public IPapyrusClass<PapyrusLeveledBase>
 {
 public:
-  PapyrusLeveledBase(const std::string& name)
-    : strName(name)
-  {
-  }
+  PapyrusLeveledBase(const std::string& name);
   const char* GetName() override { return strName.c_str(); }
 
   VarValue GetNthForm(VarValue self, const std::vector<VarValue>& arguments);
