+++ skymp5-server/cpp/server_guest_lib/script_classes/PapyrusSound.cpp
@@ -1,6 +1,7 @@
 #include "PapyrusSound.h"
 
 #include "MpObjectReference.h"
+#include "SpSnippetFunctionGen.h"
 #include "script_objects/EspmGameObject.h"
 #include "script_objects/MpFormGameObject.h"
